modif uint16 en int

+++ symboltable.c
@@ -18,7 +18,7 @@ type_table stack = {0};
 
 
 // add the symbol to the table with id, depth and whether it is a constant or not
-int add_symbol(const uint16_t id, int cons, uint16_t depth)
+int add_symbol(const int id, int cons, int depth)
 {
     type_symbol * new_symbol = &(stack.entries[stack.pointer]);
     strcpy(new_symbol->id, id);
@@ -32,7 +32,7 @@ int add_symbol(const uint16_t id, int cons, uint16_t depth)
 }
 
 // return the address of the symbol (id)
-int get_address(const uint16_t id, uint16_t depth)
+int get_address(const int id, int depth)
 {
     for(int i=0; i<stack.pointer; i++)
     {
@@ -47,13 +47,13 @@ int get_address(const uint16_t id, uint16_t depth)
 }
 
 // check whether the symbol is initialized
-int is_initialized(const uint16_t id, uint16_t depth)
+int is_initialized(const int id, int depth)
 {
     int addr = get_address(id, depth);
     return stack.entries[addr].init;
 }
 // check whether the symbol is a constant
-int is_constant(const uint16_t id, uint16_t depth)
+int is_constant(const int id, int depth)
 {
     int addr = get_address(id, depth);
     return stack.entries[addr].constant;
@@ -64,7 +64,7 @@ int get_last_pointer(){
     return stack.pointer-1;
 }
 
-void set_initialized(const uint16_t id, uint16_t depth)
+void set_initialized(const int id, int depth)
 {
     int addr = get_address(id, depth);
     stack.entries[addr].init = 1;
