+++ symboltable.h
@@ -5,20 +5,20 @@
 
 struct struct_symbol
 {
-    uint16_t id;
+    int id;
     int constant;
     int init;
-    uint16_t depth;
+    int depth;
 };
 
 typedef struct struct_symbol type_symbol;
 
-int add_symbol(const uint16_t, int, uint16_t);
-int get_address(const uint16_t, uint16_t);
-int is_initialized(const uint16_t, uint16_t);
-int is_constant(const uint16_t, uint16_t);
+int add_symbol(const int, int, int);
+int get_address(const int, int);
+int is_initialized(const int, int);
+int is_constant(const int, int);
 int get_last_pointer();
 void pop_symbol();
-void set_initialized(const uint16_t, uint16_t);
+void set_initialized(const int, int);
 
 #endif
